SolicitudEspecial y solicitud rapida

+++ database/migrations/2024_04_25_010913_administrador.php
@@ -10,7 +10,7 @@ class Administrador extends Migration
     {
         Schema::create('administrador', function (Blueprint $table) {
             $table->id('idadministrador');
-            $table->integer('nombreadministrador')->nullable();
+            $table->unsignedBigInteger('nombreadministrador')->nullable();
         });
     }
